# api/openapi.yaml
openapi: 3.1.0
info:
  title: CAGE - Contained AI-generated Code Execution
  description: |
    A secure sandboxed code execution platform for LLM agents.

    CAGE provides isolated container environments for safely executing
    LLM-generated code with fine-grained access control, resource limits,
    and comprehensive auditing.
  version: 1.0.0
  contact:
    name: CAGE Support
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://cage.example.com
    description: Production server
...
paths:
...
  /api/v1/jobs/{job_id}:
    get:
      tags:
        - execution
      summary: Get async job status and result
      description: Returns the status and result (if complete) of an async execution job
      operationId: getJobStatus
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: The job ID returned from async execute
      responses:
        '200':
          description: Job status retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/JobStatusResponse'
        '404':
          description: Job not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/files:
    get:
      tags:
        - files
      summary: List files in workspace
      description: Returns a list of files in the user's sandbox workspace directory
      operationId: listFiles
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: "/"
          description: Subdirectory path within workspace (default is root)
        - name: recursive
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: List files recursively
      responses:
        '200':
          description: File list retrieved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileListResponse'
        '401':
          description: Authentication required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      tags:
        - files
      summary: Upload file to workspace
      description: |
        Uploads a file to the user's sandbox workspace.
        Maximum file size is configurable (default 100MB).
      operationId: uploadFile
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - file
              properties:
                file:
                  type: string
                  format: binary
                  description: File to upload
                path:
                  type: string
                  description: Target path within workspace (default is root)
                  default: "/"
          application/json:
            schema:
              $ref: '#/components/schemas/FileUploadRequest'
      responses:
        '201':
          description: File uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FileUploadResponse'
        '400':
          description: Invalid file or path
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '413':
          description: File too large
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/files/{filepath}:
    get:
      tags:
        - files
      summary: Download file from workspace
      description: Downloads a file from the user's sandbox workspace
      operationId: downloadFile
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: filepath
          in: path
          required: true
          schema:
            type: string
          description: Path to file within workspace (URL encoded)
      responses:
        '200':
          description: File content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            text/plain:
              schema:
                type: string
            image/*:
              schema:
                type: string
                format: binary
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags:
        - files
      summary: Delete file from workspace
      description: Deletes a file from the user's sandbox workspace
      operationId: deleteFile
      security:
        - bearerAuth: []
        - apiKeyAuth: []
      parameters:
        - name: filepath
          in: path
          required: true
          schema:
            type: string
          description: Path to file within workspace (URL encoded)
      responses:
        '204':
          description: File deleted successfully
        '404':
          description: File not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: User JWT token for authentication
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service authentication
    adminAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Admin JWT token with elevated privileges
  schemas:
...
    ExecuteResponse:
      type: object
      required:
        - execution_id
        - status
      properties:
        execution_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [success, error, timeout, killed]
        stdout:
          type: string
          description: Standard output from execution
        stderr:
          type: string
          description: Standard error from execution
        exit_code:
          type: integer
          description: Process exit code
        duration_ms:
          type: integer
          format: int64
          description: Execution duration in milliseconds
        files_created:
          type: array
          items:
            type: string
          description: List of files created during execution
        resource_usage:
          $ref: '#/components/schemas/ResourceUsage'
...
    JobStatusResponse:
      type: object
      required:
        - job_id
        - status
      properties:
        job_id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed, timeout]
        result:
          $ref: '#/components/schemas/ExecuteResponse'
        queued_at:
          type: string
          format: date-time
        started_at:
          type: string
          format: date-time
        completed_at:
          type: string
          format: date-time
    FileListResponse:
      type: object
      required:
        - files
        - path
      properties:
        path:
          type: string
          description: Current directory path
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileInfo'
        total_size_bytes:
          type: integer
          format: int64
    FileInfo:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        path:
          type: string
        type:
          type: string
          enum: [file, directory]
        size_bytes:
          type: integer
          format: int64
        modified_at:
          type: string
          format: date-time
        permissions:
          type: string
          example: "rw-r--r--"
    FileUploadRequest:
      type: object
      required:
        - filename
        - content
      properties:
        filename:
          type: string
          description: Target filename
          pattern: "^[a-zA-Z0-9._-]+$"
        path:
          type: string
          description: Target directory path
          default: "/"
        content:
          type: string
          format: byte
          description: Base64 encoded file content
        overwrite:
          type: boolean
          default: true
          description: Overwrite if file exists
    FileUploadResponse:
      type: object
      required:
        - path
        - size_bytes
      properties:
        path:
          type: string
          description: Full path of uploaded file
        size_bytes:
          type: integer
          format: int64
        checksum:
          type: string
          description: SHA-256 checksum of uploaded file
...
    ResourceUsage:
      type: object
      properties:
        memory_mb:
          type: number
          format: float
        cpu_percent:
          type: number
          format: float
        disk_mb:
          type: number
          format: float
        pids:
          type: integer
...
    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code
          example: "EXECUTION_TIMEOUT"
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        request_id:
          type: string
          format: uuid
          description: Request ID for debugging